Перенесла вспомогательные функции в util.js
